股票分析助手/requirements.txt
# 核心依赖
langgraph>=0.0.50
langchain-core>=0.1.0
langchain-openai>=0.0.5

# 向量检索
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# 工具依赖
tavily-python>=0.3.0

# 环境变量
python-dotenv>=1.0.0

# 数据处理
numpy>=1.24.0

# 开发依赖（可选）
# pytest>=7.0.0
# black>=23.0.0
